split out checkpoint functions into separate files
