+++ SerialPrograms/CMakeLists.txt
@@ -1426,6 +1426,18 @@ file(GLOB MAIN_SOURCES
     Source/PokemonSV/PokemonSV_Panels.h
     Source/PokemonSV/PokemonSV_Settings.cpp
     Source/PokemonSV/PokemonSV_Settings.h
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_00_04.cpp
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_00_04.h   
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_05_10.cpp
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_05_10.h
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_11_15.cpp
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_11_15.h
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_16_20.cpp
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_16_20.h     
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_21_25.cpp
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_21_25.h
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_26_30.cpp
+    Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_26_30.h       
     Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory.cpp
     Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory.h
     Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStoryTools.cpp
