+++ SerialPrograms/Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStoryTools.cpp
@@ -656,1639 +656,7 @@ void checkpoint_save(SingleSwitchProgramEnvironment& env, BotBaseContext& contex
     send_program_status_notification(env, notif_status_update, "Saved at checkpoint.");        
 }
 
-void checkpoint_00(SingleSwitchProgramEnvironment& env, BotBaseContext& context){
 
-
-    // Mash A through intro cutscene, until the L stick button is detected
-    WhiteButtonWatcher leftstick(COLOR_GREEN, WhiteButton::ButtonLStick, {0.435, 0.912, 0.046, 0.047});
-    context.wait_for_all_requests();
-    run_until(
-        env.console, context,
-        [](BotBaseContext& context){
-            pbf_mash_button(context, BUTTON_A, 240 * TICKS_PER_SECOND);
-        },
-        {leftstick}
-    );
-    
-    // Stand up from chair and walk to left side of room
-    pbf_move_left_joystick(context, 128, 255, 3 * TICKS_PER_SECOND, 5 * TICKS_PER_SECOND);
-    pbf_move_left_joystick(context,   0, 128, 6 * TICKS_PER_SECOND, 1 * TICKS_PER_SECOND);
-
-}
-
-void checkpoint_01(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update, 
-    Language language
-){
-AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if(first_attempt){
-            save_game_tutorial(env.program_info(), env.console, context);
-            stats.m_checkpoint++;
-            env.update_stats();
-            send_program_status_notification(env, notif_status_update, "Saved at checkpoint.");     
-            first_attempt = false;
-        }
-        
-        context.wait_for_all_requests();
-        // set settings
-        enter_menu_from_overworld(env.program_info(), env.console, context, 0, MenuSide::RIGHT, false);
-        change_settings(env, context, language, first_attempt);
-        pbf_mash_button(context, BUTTON_B, 2 * TICKS_PER_SECOND);
-
-        break;  
-    }catch(...){
-        // (void)e;
-        first_attempt = false;
-        context.wait_for_all_requests();
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }
-    }
-}
-
-void checkpoint_02(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if(!first_attempt){
-            save_game_tutorial(env.program_info(), env.console, context);
-            stats.m_checkpoint++;
-            env.update_stats();
-            send_program_status_notification(env, notif_status_update, "Saved at checkpoint.");     
-            first_attempt = false;
-        }
-        
-        context.wait_for_all_requests();
-        env.console.log("Go downstairs, get stopped by Skwovet");
-        env.console.overlay().add_log("Go downstairs, get stopped by Skwovet", COLOR_WHITE);
-        pbf_move_left_joystick(context, 128,   0, 3 * TICKS_PER_SECOND, 20);
-        pbf_move_left_joystick(context,   0, 128, 3 * TICKS_PER_SECOND, 20);
-        pbf_move_left_joystick(context, 128, 255, 3 * TICKS_PER_SECOND, 20);
-        pbf_wait(context, 5 * TICKS_PER_SECOND);
-        // clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 5, {});
-
-        context.wait_for_all_requests();
-        env.console.log("Go to the kitchen, talk with mom");
-        env.console.overlay().add_log("Go to the kitchen, talk with mom", COLOR_WHITE);
-        pbf_move_left_joystick(context, 128, 255, 2 * TICKS_PER_SECOND, 20);
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 0, 128);
-
-        env.console.log("clear_dialog: Talk with Mom.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 5, {});
-
-        context.wait_for_all_requests();
-        env.console.log("Go to the front door, talk with Clavell");
-        env.console.overlay().add_log("Go to the front door, talk with Clavell", COLOR_WHITE);
-        pbf_move_left_joystick(context, 230, 200, 2 * TICKS_PER_SECOND, 20);
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 255, 128);
-        
-        env.console.log("clear_dialog: Talk with Clavell at front door.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 5, {});
-
-        context.wait_for_all_requests();
-        env.console.log("Go upstairs, dress up");
-        env.console.overlay().add_log("Go upstairs, dress up", COLOR_WHITE);
-        pbf_move_left_joystick(context, 255, 128, 2 * TICKS_PER_SECOND, 20);
-        pbf_move_left_joystick(context, 185,  10, 1 * TICKS_PER_SECOND, 20);
-        pbf_move_left_joystick(context, 128,   0, 4 * TICKS_PER_SECOND, 20);
-        pbf_move_left_joystick(context, 255, 128, 4 * TICKS_PER_SECOND, 20);
-        pbf_move_left_joystick(context, 110, 200, 3 * TICKS_PER_SECOND, 20);
-        pbf_move_left_joystick(context, 255, 128, 2 * TICKS_PER_SECOND, 20);
-        pbf_mash_button(context, BUTTON_A, 20 * TICKS_PER_SECOND);
-
-        context.wait_for_all_requests();
-        env.console.log("Go to the living room, talk with Clavell");
-        env.console.overlay().add_log("Go to the living room, talk with Clavell", COLOR_WHITE);
-        pbf_move_left_joystick(context,   0,   0, 3 * TICKS_PER_SECOND, 20);
-        pbf_move_left_joystick(context,   0, 128, 3 * TICKS_PER_SECOND, 20);
-        pbf_move_left_joystick(context, 128, 255, 4 * TICKS_PER_SECOND, 20);
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 0, 128);
-        
-        env.console.log("clear_dialog: Talk with Clavell at living room.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 10, {});
-
-        context.wait_for_all_requests();
-        env.console.log("Go outside, receive Rotom Phone");
-        env.console.overlay().add_log("Go outside, receive Rotom Phone", COLOR_WHITE);
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 245, 230);
-
-        env.console.log("clear_dialog: Talk with Clavell outside. Receive Rotom phone. Stop when detect overworld.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_OVERWORLD, 60, {ClearDialogCallback::OVERWORLD, ClearDialogCallback::WHITE_A_BUTTON});
-
-        context.wait_for_all_requests();
-        env.console.log("Bump into power of science NPC");
-        // console.overlay().add_log("Bump into power of science NPC", COLOR_WHITE);
-        pbf_move_left_joystick(context, 128,   0, 33 * TICKS_PER_SECOND, 20);
-
-        context.wait_for_all_requests();
-        env.console.log("Clear map tutorial");
-        // console.overlay().add_log("Clear map tutorial", COLOR_WHITE);
-        open_map_from_overworld(env.program_info(), env.console, context, true);
-        leave_phone_to_overworld(env.program_info(), env.console, context);
-
-        break;  
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }
-    }
-}
-
-void checkpoint_03(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update,
-    Language language,
-    StarterChoice starter_choice
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){   
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }
-        
-        context.wait_for_all_requests();
-        
-        pbf_move_left_joystick(context, 255, 0, 1 * TICKS_PER_SECOND, 20);
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 255, 156, 1 * TICKS_PER_SECOND);
-        env.console.log("overworld_navigation(): Go to Nemona's house.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 128, 0);
-        
-        context.wait_for_all_requests();
-        env.console.log("Entered Nemona's house");
-        mash_button_till_overworld(env.console, context);
-        context.wait_for_all_requests();
-        env.console.log("Picking a starter...");
-        switch(starter_choice){
-        case StarterChoice::SPRIGATITO:
-            env.console.log("Picking Sprigatito...");
-            pbf_move_left_joystick(context, 75, 0, 80, 20);
-            break;
-        case StarterChoice::FUECOCO:
-            env.console.log("Picking Fuecoco...");
-            pbf_move_left_joystick(context, 180, 0, 80, 20);
-            break;
-        case StarterChoice::QUAXLY:
-            env.console.log("Picking Quaxly...");
-            pbf_move_left_joystick(context, 128, 0, 80, 20);
-            break;
-        }
-        pbf_press_button(context, BUTTON_A, 20, 105); // choose the starter
-        env.console.log("clear_dialog: Choose starter. Stop when detect prompt to receive starter.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_PROMPT, 20, {ClearDialogCallback::PROMPT_DIALOG});
-
-        pbf_press_button(context, BUTTON_A, 20, 105); // accept the pokemon
-        env.console.log("clear_dialog: Stop when detect prompt to give nickname to starter.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_PROMPT, 20, {ClearDialogCallback::PROMPT_DIALOG});
-
-        pbf_mash_button(context, BUTTON_B, 100);  // Don't give a nickname
-        env.console.log("clear_dialog: Talk to Nemona and Clavell. Stop when detect overworld.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_OVERWORLD, 20, {ClearDialogCallback::OVERWORLD});
-        
-        context.wait_for_all_requests();
-        env.console.log("Clear auto heal tutorial.");
-        // Press X until Auto heal tutorial shows up
-        TutorialWatcher tutorial;
-        int ret = run_until(
-            env.console, context,
-            [](BotBaseContext& context){
-                for (int i = 0; i < 10; i++){
-                    pbf_press_button(context, BUTTON_X, 20, 250);
-                }
-            },
-            {tutorial}
-        );
-        if (ret < 0){
-            throw OperationFailedException(
-                ErrorReport::SEND_ERROR_REPORT, env.console,
-                "Stuck trying to clear auto heal tutorial.",
-                true
-            );  
-        }
-        clear_tutorial(env.console, context);
-
-        env.console.log("Change move order.");
-        swap_starter_moves(env.program_info(), env.console, context, language);
-        leave_box_system_to_overworld(env.program_info(), env.console, context);
-
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }
-    }
-
-}
-
-void checkpoint_04(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{        
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }        
-        context.wait_for_all_requests();
-
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 220, 245, 50);
-        pbf_move_left_joystick(context, 128, 0, 4 * TICKS_PER_SECOND, 1 * TICKS_PER_SECOND);
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 255, 128, 50);
-        pbf_move_left_joystick(context, 128, 0, 4 * TICKS_PER_SECOND, 1 * TICKS_PER_SECOND);
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 255, 60, 50);
-        pbf_move_left_joystick(context, 128, 0, 4 * TICKS_PER_SECOND, 2 * TICKS_PER_SECOND);
-        env.console.log("overworld_navigation: Go to Nemona at the beach.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_SPAM_A, 128, 0, 8);
-        
-        context.wait_for_all_requests();
-        env.console.overlay().add_log("Found Nemona", COLOR_WHITE);
-
-        context.wait_for_all_requests();
-        env.console.log("Starting battle...");
-        env.console.overlay().add_log("Starting battle...", COLOR_WHITE);
-        // TODO: Battle start prompt detection
-        // can lose this battle, and story will continue
-        mash_button_till_overworld(env.console, context);
-        context.wait_for_all_requests();
-        env.console.log("Finished battle.");
-        env.console.overlay().add_log("Finished battle.", COLOR_WHITE);        
-
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }        
-    }
-
-}
-
-void checkpoint_05(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 40, 160, 60);
-        pbf_move_left_joystick(context, 128, 0, 7 * TICKS_PER_SECOND, 20);
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 40, 84, 60);
-        env.console.log("overworld_navigation: Go to mom at the gate.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 128, 0, 20);
-        
-        context.wait_for_all_requests();
-        env.console.log("Get mom's sandwich");
-        env.console.overlay().add_log("Get mom's sandwich", COLOR_WHITE);
-        mash_button_till_overworld(env.console, context);
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }    
-}
-
-void checkpoint_06(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{        
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }        
-
-        context.wait_for_all_requests();
-
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 40, 82, 60);
-        pbf_move_left_joystick(context, 128, 0, 6 * TICKS_PER_SECOND, 20);
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 110, 10, 60);
-        env.console.log("overworld_navigation: Go to Nemona.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 128, 0, 20);
-        
-        context.wait_for_all_requests();
-        env.console.log("clear_dialog: Talk with Nemona to start catch tutorial. Stop when detect battle.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_BATTLE, 60, 
-            {ClearDialogCallback::WHITE_A_BUTTON, ClearDialogCallback::TUTORIAL, ClearDialogCallback::BATTLE});
-        
-        // can die in catch tutorial, and the story will continue
-        env.console.log("run_battle_press_A: Battle Lechonk in catch tutorial. Stop when detect dialog.");
-        run_battle_press_A(env.console, context, BattleStopCondition::STOP_DIALOG);
-
-        env.console.log("clear_dialog: Talk with Nemona to finish catch tutorial. Stop when detect overworld.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_OVERWORLD, 60, 
-            {ClearDialogCallback::TUTORIAL, ClearDialogCallback::OVERWORLD});
-
-        context.wait_for_all_requests();
-        env.console.log("Finished catch tutorial");
-        env.console.overlay().add_log("Finished catch tutorial", COLOR_WHITE);
-
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }            
-    }
-}
-
-void checkpoint_07(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-
-        context.wait_for_all_requests();
-        env.console.log("Move to cliff");
-        env.console.overlay().add_log("Move to cliff", COLOR_WHITE);
-
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 240, 60, 80);
-        env.console.log("overworld_navigation: Go to cliff.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 116, 0, 72, 24);
-
-        env.console.log("clear_dialog: Talk to Nemona at the cliff. Stop when detect overworld.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_OVERWORLD, 60, {ClearDialogCallback::OVERWORLD});
-
-        context.wait_for_all_requests();
-        env.console.log("Mystery cry");
-        env.console.overlay().add_log("Mystery cry", COLOR_WHITE);
-        
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }           
-    }
-}
-
-void checkpoint_08(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 230, 70, 100);
-        env.console.log("overworld_navigation: Go to cliff.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 128, 0);
-
-        env.console.log("clear_dialog: Look over the injured Miraidon/Koraidon on the beach. Fall down the cliff");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 30, {});
-        // long animation
-        env.console.log("overworld_navigation: Go to Legendary pokemon laying on the beach.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 128, 0, 30);
-
-        env.console.log("clear_dialog: Offer Miraidon/Koraidon a sandwich.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 10, {});
-
-        // TODO: Bag menu navigation
-        context.wait_for_all_requests();
-        env.console.log("Feed mom's sandwich");
-        env.console.overlay().add_log("Feed mom's sandwich", COLOR_WHITE);
-        
-        GradientArrowWatcher arrow(COLOR_RED, GradientArrowType::RIGHT, {0.104, 0.312, 0.043, 0.08});
-        context.wait_for_all_requests();
-
-        int ret = run_until(
-            env.console, context,
-            [](BotBaseContext& context){
-                for (int i = 0; i < 10; i++){
-                    pbf_press_dpad(context, DPAD_UP, 20, 250);
-                }
-            },
-            {arrow}
-        );
-        if (ret < 0){
-            throw OperationFailedException(
-                ErrorReport::SEND_ERROR_REPORT, env.console,
-                "Failed to feed mom's sandwich.",
-                true
-            );  
-        }
-
-        // only press A when the sandwich is selected
-        pbf_mash_button(context, BUTTON_A, 100);
-
-        env.console.log("clear_dialog: Miraidon/Koraidon gets up and walks to cave entrance.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 35, {});
-        // long animation
-
-        // First Nemona cave conversation
-        context.wait_for_all_requests();
-        env.console.log("Enter cave");
-        env.console.overlay().add_log("Enter cave", COLOR_WHITE);
-        do_action_and_monitor_for_battles(env, env.console, context,
-            [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){
-                pbf_move_left_joystick(context, 128, 20, 10 * TICKS_PER_SECOND, 20);
-                pbf_move_left_joystick(context, 150, 20, 1 * TICKS_PER_SECOND, 20);
-                pbf_move_left_joystick(context, 128, 20, 8 * TICKS_PER_SECOND, 20);
-                pbf_move_left_joystick(context, 150, 20, 2 * TICKS_PER_SECOND, 20);                
-            }
-        );
-        
-        env.console.log("overworld_navigation: Go to cave.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 128, 20, 10);
-
-        env.console.log("clear_dialog: Talk to Nemona yelling down, while you're down in the cave.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 10, {ClearDialogCallback::PROMPT_DIALOG});
-
-        do_action_and_monitor_for_battles(env, env.console, context,
-            [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){
-                // Legendary rock break
-                context.wait_for_all_requests();
-                console.log("Rock break");
-                console.overlay().add_log("Rock break", COLOR_WHITE);
-                pbf_move_left_joystick(context, 128, 20, 3 * TICKS_PER_SECOND, 20);
-                realign_player(env.program_info(), console, context, PlayerRealignMode::REALIGN_NO_MARKER, 230, 25, 30);
-                pbf_move_left_joystick(context, 128, 0, 2 * TICKS_PER_SECOND, 5 * TICKS_PER_SECOND);
-
-                // Houndour wave
-                context.wait_for_all_requests();
-                console.log("Houndour wave");
-                console.overlay().add_log("Houndour wave", COLOR_WHITE);
-                pbf_move_left_joystick(context, 140, 20, 4 * TICKS_PER_SECOND, 2 * TICKS_PER_SECOND);
-                realign_player(env.program_info(), console, context, PlayerRealignMode::REALIGN_NO_MARKER, 220, 15, 30);
-                pbf_move_left_joystick(context, 128, 20, 5 * TICKS_PER_SECOND, 2 * TICKS_PER_SECOND);
-                pbf_move_left_joystick(context, 128, 20, 6 * TICKS_PER_SECOND, 2 * TICKS_PER_SECOND);
-                realign_player(env.program_info(), console, context, PlayerRealignMode::REALIGN_NO_MARKER, 220, 25, 20);
-                pbf_move_left_joystick(context, 128, 20, 4 * TICKS_PER_SECOND, 2 * TICKS_PER_SECOND);
-                pbf_move_left_joystick(context, 128, 20, 4 * TICKS_PER_SECOND, 2 * TICKS_PER_SECOND);
-                realign_player(env.program_info(), console, context, PlayerRealignMode::REALIGN_NO_MARKER, 220, 25, 25);
-                pbf_move_left_joystick(context, 128, 20, 6 * TICKS_PER_SECOND, 20 * TICKS_PER_SECOND);
-
-                // Houndoom encounter
-                context.wait_for_all_requests();
-                console.log("Houndoom encounter");
-                console.overlay().add_log("Houndoom encounter", COLOR_WHITE);
-                pbf_move_left_joystick(context, 128, 20, 4 * TICKS_PER_SECOND, 20);
-                realign_player(env.program_info(), console, context, PlayerRealignMode::REALIGN_NO_MARKER, 245, 20, 20);
-                pbf_move_left_joystick(context, 128, 20, 2 * TICKS_PER_SECOND, 20);
-                realign_player(env.program_info(), console, context, PlayerRealignMode::REALIGN_NO_MARKER, 255, 90, 20);
-                pbf_move_left_joystick(context, 128, 20, 8 * TICKS_PER_SECOND, 8 * TICKS_PER_SECOND);
-                pbf_press_button(context, BUTTON_L, 20, 20);
-            }
-        );
-        
-        env.console.log("overworld_navigation: Go to Houndoom.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 128, 20, 40);
-        
-        mash_button_till_overworld(env.console, context, BUTTON_A);
-
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }            
-    }
-
-}
-
-
-void checkpoint_09(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{        
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }      
-        context.wait_for_all_requests();
-
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 230, 120, 100);
-        env.console.log("overworld_navigation: Go to Arven at the tower.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 128, 0);
-            
-        context.wait_for_all_requests();
-        env.console.log("Found Arven");
-        env.console.overlay().add_log("Found Arven", COLOR_WHITE);
-        // can lose battle, and story will still continue
-        mash_button_till_overworld(env.console, context, BUTTON_A);
-        context.wait_for_all_requests();
-        env.console.log("Receive legendary ball");
-        env.console.overlay().add_log("Receive legendary ball", COLOR_WHITE);
-
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }            
-    }
-}
-
-void checkpoint_10(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }        
-        context.wait_for_all_requests();
-        env.console.log("Lighthouse view");
-        env.console.overlay().add_log("Lighthouse view", COLOR_WHITE);
-        do_action_and_monitor_for_battles(env, env.console, context,
-            [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){
-                realign_player(env.program_info(), console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 230, 110, 100);
-                pbf_move_left_joystick(context, 128, 0, 6 * TICKS_PER_SECOND, 8 * TICKS_PER_SECOND);
-                pbf_move_left_joystick(context, 128, 0, 4 * TICKS_PER_SECOND, 20);
-                pbf_move_left_joystick(context, 255, 128, 15, 20);
-                pbf_press_button(context, BUTTON_L, 20, 20);
-                pbf_move_left_joystick(context, 128, 0, 7 * TICKS_PER_SECOND, 20);                
-            }
-        );
-
-        env.console.log("overworld_navigation: Go to Nemona on the lighthouse.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_SPAM_A, 128, 0, 20);
-
-        mash_button_till_overworld(env.console, context, BUTTON_A);
-
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }           
-    }
-}
-
-void checkpoint_11(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        } 
-
-        context.wait_for_all_requests();
-        do_action_and_monitor_for_battles(env, env.console, context,
-            [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){
-                realign_player(env.program_info(), console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 100, 210, 100);
-                pbf_move_left_joystick(context, 128, 0, 187, 20);
-                pbf_move_left_joystick(context, 0, 128, 30, 8 * TICKS_PER_SECOND);
-                pbf_move_left_joystick(context, 128, 0, 1 * TICKS_PER_SECOND, 2 * TICKS_PER_SECOND);
-
-                realign_player(env.program_info(), console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 100, 60, 200);                
-            }
-        );     
-
-        env.console.log("overworld_navigation: Go to Los Platos.");
-        overworld_navigation(env.program_info(), env.console, context, NavigationStopCondition::STOP_DIALOG, NavigationMovementMode::DIRECTIONAL_ONLY, 128, 0, 75);
-
-        env.console.log("clear_dialog: Talk with Nemona at Los Platos. Clear Let's go tutorial. Stop when detect overworld.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_OVERWORLD, 60, {ClearDialogCallback::TUTORIAL, ClearDialogCallback::OVERWORLD});
-
-        context.wait_for_all_requests();
-
-        env.console.log("Reached Los Platos");
-        env.console.overlay().add_log("Reached Los Platos", COLOR_WHITE);
-
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-
-void checkpoint_12(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    // reset rate: ~25%. 12 resets out of 52. 
-    // resets due to: getting attacked by wild pokemon, either from behind, 
-    // or when lead pokemon not strong enough to clear them with Let's go
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        do_action_and_monitor_for_battles(env, env.console, context,
-        [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){        
-        
-            if (first_attempt){
-                checkpoint_save(env, context, notif_status_update);
-                first_attempt = false;
-            } 
-
-            context.wait_for_all_requests();
-            
-            fly_to_overlapping_flypoint(env.program_info(), env.console, context);
-
-            // re-orient camera
-            pbf_press_button(context, BUTTON_L, 20, 20);
-
-            walk_forward_while_clear_front_path(env.program_info(), env.console, context, 35);
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 128, 0, 100);
-            walk_forward_while_clear_front_path(env.program_info(), env.console, context, 500);
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 100, 0, 120);
-            walk_forward_while_clear_front_path(env.program_info(), env.console, context, 2000);
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 255, 0, 120);
-            walk_forward_while_clear_front_path(env.program_info(), env.console, context, 1250);
-            
-            // check we're not still at the Los Platos Pokecenter.
-            confirm_no_overlapping_flypoint(env.program_info(), env.console, context);
-
-            // not stuck at Los Platos Pokecenter
-            pbf_press_button(context, BUTTON_B, 20, 1 * TICKS_PER_SECOND);
-            pbf_press_button(context, BUTTON_B, 20, 1 * TICKS_PER_SECOND);
-            press_Bs_to_back_to_overworld(env.program_info(), env.console, context, 7);
-
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 80, 0, 100);
-            walk_forward_while_clear_front_path(env.program_info(), env.console, context, 1500);
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 30, 0, 50);
-            walk_forward_while_clear_front_path(env.program_info(), env.console, context, 1000);
-
-            heal_at_pokecenter(env.program_info(), env.console, context);
-   
-        }  
-        );
-
-        env.console.log("Reached Mesagoza (South) Pokecenter.");
-        
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_13(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    // reset rate: 0%. 0 resets out of 70.
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        do_action_and_monitor_for_battles(env, env.console, context,
-        [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){        
-        
-            if (first_attempt){
-                checkpoint_save(env, context, notif_status_update);
-                first_attempt = false;
-            } 
-
-            fly_to_overlapping_flypoint(env.program_info(), env.console, context);
-
-            context.wait_for_all_requests();
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 0, 80, 50);
-            walk_forward_while_clear_front_path(env.program_info(), env.console, context, 500);
-            walk_forward_until_dialog(env.program_info(), env.console, context, NavigationMovementMode::DIRECTIONAL_ONLY, 30);
-        });
-
-        env.console.log("clear_dialog: Talk with Nemona at Mesagoza gate. Stop when detect battle.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_BATTLE, 60,
-            {ClearDialogCallback::PROMPT_DIALOG, ClearDialogCallback::DIALOG_ARROW, ClearDialogCallback::BATTLE});
-        
-        env.console.log("run_battle_press_A: Battle with Nemona at Mesagoza gate. Stop when detect dialog.");
-        run_battle_press_A(env.console, context, BattleStopCondition::STOP_DIALOG);
-        
-        env.console.log("clear_dialog: Talk with Nemona within Mesagoza. Stop when detect overworld.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_OVERWORLD, 60, 
-            {ClearDialogCallback::OVERWORLD, ClearDialogCallback::PROMPT_DIALOG, ClearDialogCallback::WHITE_A_BUTTON});
-        
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_14(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-        // realign diagonally to the left
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 80, 0, 100);
-        // walk forward so you're off center
-        pbf_move_left_joystick(context, 128, 0, 100, 100);
-        // realign going straight
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 128, 0, 100);
-        // walk forward, while still off center
-        pbf_move_left_joystick(context, 128, 0, 2000, 100);
-        // realign diagonally to the right
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 178, 0, 100);
-        // walk forward so you're closer to the center
-        pbf_move_left_joystick(context, 128, 0, 150, 100);
-        // realign going straight
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 128, 0, 100);
-        // walk forward until hit dialog at top of stairs
-        walk_forward_until_dialog(env.program_info(), env.console, context, NavigationMovementMode::DIRECTIONAL_ONLY, 60);
-        // clear dialog until battle. with prompt, battle
-        env.console.log("clear_dialog: Talk with Team Star at the top of the stairs. Stop when detect battle.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_BATTLE, 60, {ClearDialogCallback::PROMPT_DIALOG, ClearDialogCallback::BATTLE, ClearDialogCallback::DIALOG_ARROW});
-        // run battle until dialog
-        env.console.log("run_battle_press_A: Battle with Team Star grunt 1. Stop when detect dialog.");
-        run_battle_press_A(env.console, context, BattleStopCondition::STOP_DIALOG);
-        // clear dialog until battle, with prompt, white button, tutorial, battle
-        env.console.log("clear_dialog: Talk with Team Star and Nemona. Receive Tera orb. Stop when detect battle.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_BATTLE, 60, 
-            {ClearDialogCallback::PROMPT_DIALOG, ClearDialogCallback::WHITE_A_BUTTON, ClearDialogCallback::TUTORIAL, ClearDialogCallback::BATTLE, ClearDialogCallback::DIALOG_ARROW});
-        // run battle until dialog
-        env.console.log("run_battle_press_A: Battle with Team Star grunt 2. Stop when detect dialog.");
-        run_battle_press_A(env.console, context, BattleStopCondition::STOP_DIALOG);
-        // clear dialog until overworld
-        clear_dialog(env.console, context, ClearDialogMode::STOP_OVERWORLD, 60, {ClearDialogCallback::OVERWORLD});
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_15(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-        // realign diagonally to the right
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 178, 0, 100);
-        // walk forward so you're closer to the center
-        pbf_move_left_joystick(context, 128, 0, 100, 100);
-        // realign going straight
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 128, 0, 100);
-        // walk forward up stairs
-        pbf_move_left_joystick(context, 128, 0, 1000, 100);
-        // realign going straight
-        realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_OLD_MARKER);
-        // walk forward until hit dialog inside the school
-        walk_forward_until_dialog(env.program_info(), env.console, context, NavigationMovementMode::DIRECTIONAL_ONLY, 60);
-
-        env.console.log("clear_dialog: Talk with Nemona, Clavell, and Jacq inside the school. Stop when detect overworld.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_OVERWORLD, 60, 
-            {ClearDialogCallback::PROMPT_DIALOG, ClearDialogCallback::OVERWORLD});
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_16(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-
-        // walk left
-        pbf_move_left_joystick(context, 0, 128, 400, 100);
-        // walk down to classroom exit.
-        pbf_move_left_joystick(context, 128, 255, 300, 100);
-        env.console.log("clear_dialog: Leave classroom.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 5);
-
-        // Wait for detection of school navigation menu
-        wait_for_gradient_arrow(env.program_info(), env.console, context, {0.031, 0.193, 0.047, 0.078}, 5);
-
-        // enter Cafeteria
-        pbf_mash_button(context, BUTTON_A, 3 * TICKS_PER_SECOND);
-        pbf_wait(context, 3 * TICKS_PER_SECOND);
-        context.wait_for_all_requests();
-
-        // walk forward
-        pbf_move_left_joystick(context, 128, 0, 300, 100);
-        // turn left 
-        pbf_move_left_joystick(context, 0, 128, 50, 100);
-
-        // talk to Arven. stop at overworld. need prompt, overworld, white button A. and book?
-        env.console.log("Talk with Arven. Receive Titan questline (Path of Legends).");
-        press_A_until_dialog(env.program_info(), env.console, context, 1);
-        mash_button_till_overworld(env.console, context, BUTTON_A, 360);
-        
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_17(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-
-        // walk backwards until dialog
-        walk_forward_until_dialog(env.program_info(), env.console, context, NavigationMovementMode::DIRECTIONAL_ONLY, 20, 255);
-        env.console.log("Talk with Cassiopeia.");
-        mash_button_till_overworld(env.console, context, BUTTON_A, 360);
-
-        // re-orient camera
-        pbf_press_button(context, BUTTON_L, 20, 100);
-        // move backwards towards front desk
-        pbf_move_left_joystick(context, 128, 255, 200, 100);
-        // re-orient camera
-        pbf_press_button(context, BUTTON_L, 20, 100);
-        // move right towards navigation kiosk
-        pbf_move_left_joystick(context, 255, 128, 100, 100);
-        // open school navigation screen
-        press_button_until_gradient_arrow(env.program_info(), env.console, context, {0.031, 0.193, 0.047, 0.078});
-        // go to staff room
-        basic_menu_navigation(env.program_info(), env.console, context, {0.031, 0.193, 0.047, 0.078}, {0.031, 0.193 + 0.074219, 0.047, 0.078}, DPAD_DOWN, 1);
-        // enter staff room
-        pbf_mash_button(context, BUTTON_A, 3 * TICKS_PER_SECOND);
-        pbf_wait(context, 3 * TICKS_PER_SECOND);
-
-        env.console.log("clear_dialog: See Geeta. Talk to Nemona. Receive Gym/Elite Four questline (Victory Road).");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_OVERWORLD, 60, 
-            {ClearDialogCallback::OVERWORLD, ClearDialogCallback::PROMPT_DIALOG});
-
-        
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_18(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-
-        // walk down
-        pbf_move_left_joystick(context, 128, 255, 200, 100);
-        // walk left towards door
-        pbf_move_left_joystick(context, 0, 128, 100, 100);
-
-        // wait for school navigation menu
-        context.wait_for_all_requests();
-        wait_for_gradient_arrow(env.program_info(), env.console, context, {0.031, 0.193, 0.047, 0.078}, 10);
-        // enter Directors office
-        pbf_mash_button(context, BUTTON_A, 6 * TICKS_PER_SECOND);
-
-        env.console.log("Talk to Clavell in his office, and the professor.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 25, 
-            {ClearDialogCallback::PROMPT_DIALOG});  // max time between dialog: 17s. set timeout to 25 seconds for buffer.
-        // mash A to get through the Random A press that you need. when the professor shows you area zero.
-        pbf_mash_button(context, BUTTON_A, 3 * TICKS_PER_SECOND);
-        clear_dialog(env.console, context, ClearDialogMode::STOP_OVERWORLD, 60, 
-            {ClearDialogCallback::OVERWORLD, ClearDialogCallback::PROMPT_DIALOG});
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_19(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-
-        // walk right 
-        pbf_move_left_joystick(context, 255, 128, 50, 100);
-        // walk down towards door
-        pbf_move_left_joystick(context, 128, 255, 200, 100);
-
-        env.console.log("Talk to Nemona and go to dorm.");
-        mash_button_till_overworld(env.console, context, BUTTON_A, 360);
-
-        // walk forward
-        pbf_move_left_joystick(context, 128, 0, 100, 100);
-        // walk left towards bed
-        pbf_move_left_joystick(context, 0, 128, 100, 100);
-
-        env.console.log("Go to bed. Time passes until treasure hunt.");
-        mash_button_till_overworld(env.console, context, BUTTON_A, 360);
-
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_20(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        //walk right towards door
-        pbf_move_left_joystick(context, 255, 128, 200, 100);
-
-        wait_for_gradient_arrow(env.program_info(), env.console, context, {0.031, 0.193, 0.047, 0.078}, 10);
-
-        env.console.log("Leave dorm for schoolyard.");
-        mash_button_till_overworld(env.console, context, BUTTON_A, 360);
-
-        walk_forward_until_dialog(env.program_info(), env.console, context, NavigationMovementMode::DIRECTIONAL_ONLY, 60, 0);
-
-        env.console.log("Talk to Nemona, Arven, Cassiopeia.");
-        clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 16, 
-            {ClearDialogCallback::PROMPT_DIALOG, ClearDialogCallback::BLACK_DIALOG_BOX}); // max time between dialog: 11
-        
-        // mash A to get through the Random A press that you need. when the Nemona shows you a Poke Gym.
-        pbf_mash_button(context, BUTTON_A, 250);
-
-        clear_dialog(env.console, context, ClearDialogMode::STOP_TIMEOUT, 10, 
-            {ClearDialogCallback::TUTORIAL}); // max time between dialog: 3
-        
-        env.console.log("Get on ride.");
-        pbf_mash_button(context, BUTTON_PLUS, 1 * TICKS_PER_SECOND);
-
-        mash_button_till_overworld(env.console, context, BUTTON_A, 360);
-
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_21(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_22(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_23(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_24(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_25(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-// todo: uncomment checkpoint_save and fly_to_overlapping_flypoint
-void checkpoint_26(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            // checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-
-        // fly_to_overlapping_flypoint(env.program_info(), env.console, context);
-
-        // align for long stretch 1, part 1
-        do_action_and_monitor_for_battles(env, env.console, context,
-        [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){     
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 80, 0, 70);
-        });
-
-        overworld_navigation(env.program_info(), env.console, context, 
-            NavigationStopCondition::STOP_MARKER, NavigationMovementMode::CLEAR_WITH_LETS_GO, 
-            128, 0, 60, 10, false);
-
-        // align for long stretch 1, part 2
-        do_action_and_monitor_for_battles(env, env.console, context,
-        [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){     
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 120, 0, 55);
-        });
-        
-        overworld_navigation(env.program_info(), env.console, context, 
-            NavigationStopCondition::STOP_MARKER, NavigationMovementMode::CLEAR_WITH_LETS_GO, 
-            128, 0, 60, 10, false);
-
-        // align for long stretch 2
-        do_action_and_monitor_for_battles(env, env.console, context,
-        [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){     
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 70, 255, 100);
-
-        });
-
-        overworld_navigation(env.program_info(), env.console, context, 
-            NavigationStopCondition::STOP_MARKER, NavigationMovementMode::CLEAR_WITH_LETS_GO, 
-            128, 0, 60, 10, false);
-
-        // align for long stretch 3, part 1
-        do_action_and_monitor_for_battles(env, env.console, context,
-        [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){     
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 120, 0, 75);
-
-        });
-        
-        overworld_navigation(env.program_info(), env.console, context, 
-            NavigationStopCondition::STOP_MARKER, NavigationMovementMode::CLEAR_WITH_LETS_GO, 
-            128, 0, 60, 10, false);
-
-        // align for long stretch 3, part 2. // todo: consider realign based on pokecenter.
-        do_action_and_monitor_for_battles(env, env.console, context,
-        [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){     
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 35, 0, 85);
-
-        });
-
-        overworld_navigation(env.program_info(), env.console, context, 
-            NavigationStopCondition::STOP_MARKER, NavigationMovementMode::CLEAR_WITH_LETS_GO, 
-            128, 0, 60, 10, false);        
-
-        // align to cross bridge
-        do_action_and_monitor_for_battles(env, env.console, context,
-        [&](SingleSwitchProgramEnvironment& env, ConsoleHandle& console, BotBaseContext& context){     
-            realign_player(env.program_info(), env.console, context, PlayerRealignMode::REALIGN_NEW_MARKER, 0, 105, 30);
-
-        });
-
-        overworld_navigation(env.program_info(), env.console, context, 
-            NavigationStopCondition::STOP_MARKER, NavigationMovementMode::CLEAR_WITH_LETS_GO, 
-            128, 0, 30, 30, false);    
-
-                       
-              
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_27(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_28(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_29(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_30(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_31(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_32(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_33(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
-
-void checkpoint_34(
-    SingleSwitchProgramEnvironment& env, 
-    BotBaseContext& context, 
-    EventNotificationOption& notif_status_update
-){
-    AutoStoryStats& stats = env.current_stats<AutoStoryStats>();
-    bool first_attempt = true;
-    while (true){
-    try{
-        if (first_attempt){
-            checkpoint_save(env, context, notif_status_update);
-            first_attempt = false;
-        }         
-        context.wait_for_all_requests();
-       
-        break;
-    }catch(OperationFailedException& e){
-        context.wait_for_all_requests();
-        env.console.log(e.m_message, COLOR_RED);
-        env.console.log("Resetting from checkpoint.");
-        reset_game(env.program_info(), env.console, context);
-        stats.m_reset++;
-        env.update_stats();
-    }             
-    }
-
-}
