+++ SerialPrograms/Source/PokemonSV/Programs/AutoStory/PokemonSV_AutoStory_Checkpoint_21_25.h
@@ -1,0 +1,44 @@
+/*  AutostoryTools
+ *
+ *  From: https://github.com/PokemonAutomation/Arduino-Source
+ *
+ */
+
+#ifndef PokemonAutomation_PokemonSV_AutoStory_Checkpoint_21_25_H
+#define PokemonAutomation_PokemonSV_AutoStory_Checkpoint_21_25_H
+
+#include "NintendoSwitch/NintendoSwitch_SingleSwitchProgram.h"
+#include "PokemonSV_AutoStoryTools.h"
+
+namespace PokemonAutomation{
+namespace NintendoSwitch{
+namespace PokemonSV{
+
+
+
+// start: (After the break, with level 100 Gardevoir. At Mesagoza West pokecenter.)
+// end: 
+void checkpoint_21(SingleSwitchProgramEnvironment& env, BotBaseContext& context, EventNotificationOption& notif_status_update);   
+
+// start: (At South Province Area Two Pokecenter. )
+// end: 
+void checkpoint_22(SingleSwitchProgramEnvironment& env, BotBaseContext& context, EventNotificationOption& notif_status_update);   
+
+// start: (At Cortondo East Pokecenter)
+// end: 
+void checkpoint_23(SingleSwitchProgramEnvironment& env, BotBaseContext& context, EventNotificationOption& notif_status_update);   
+
+// start: (Done Cortondo Gym challenge)
+// end: 
+void checkpoint_24(SingleSwitchProgramEnvironment& env, BotBaseContext& context, EventNotificationOption& notif_status_update);   
+
+// start: (Beat Cortondo Gym leader (Bug))
+// end: 
+void checkpoint_25(SingleSwitchProgramEnvironment& env, BotBaseContext& context, EventNotificationOption& notif_status_update);  
+
+
+
+}
+}
+}
+#endif
